+++ src/engine/Color.h
@@ -8,25 +8,25 @@
 class Color {
 public:
     Color();
-    Color(float r, float g, float b, float a);
-    explicit Color(const glm::vec4 &v);
+    Color(double r, double g, double b, double a);
+    explicit Color(const glm::dvec4 &v);
 
-    void setRGB(float r, float g, float b);
-    void setRGBA(float r, float g, float b, float a);
+    void setRGB(double r, double g, double b);
+    void setRGBA(double r, double g, double b, double a);
     void copy(const Color &color);
     void setAlpha(uint8_t alpha);
     [[nodiscard]] uint32_t getUint32() const;
 
-    Color operator*(float intensity) const;
+    Color operator*(double intensity) const;
     Color& operator*=(const Color& rhs);
     Color operator+(const Color &other) const;
     Color& operator+=(const Color& rhs);
-    Color operator+(const float value) const;
+    Color operator+(const double value) const;
 private:
-    float m_red { 0 }, m_green { 0 }, m_blue { 0 }, m_alpha { 1 };
+    double m_red { 0 }, m_green { 0 }, m_blue { 0 }, m_alpha { 1 };
 
     std::array<uint8_t, 4> toRGB() const;
-    std::array<float, 4> fromRGB() const;
+    std::array<double, 4> fromRGB() const;
 };
