+++ src/raytracer/Transformations.h
@@ -9,12 +9,12 @@ public:
     Transformations(const Transformations &other) = delete;
     Transformations(const Transformations &&other) = delete;
 
-    static glm::mat4 identity();
-    static glm::mat4 translate(float x, float y, float z);
-    static glm::mat4 translate(const glm::vec3 &factor);
-    static glm::mat4 scale(float x, float y, float z);
-    static glm::mat4 scale(const glm::vec3 &factor);
-    static glm::mat4 rotate(const float angle, const glm::vec3 &axes);
+    static glm::dmat4 identity();
+    static glm::dmat4 translate(double x, double y, double z);
+    static glm::dmat4 translate(const glm::dvec3 &factor);
+    static glm::dmat4 scale(double x, double y, double z);
+    static glm::dmat4 scale(const glm::dvec3 &factor);
+    static glm::dmat4 rotate(const double angle, const glm::dvec3 &axes);
 private:
 };
